+++ Frontend/src/screens/onboarding/Onboarding2.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { View, Text, Image, TouchableOpacity } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import Icon from "react-native-vector-icons/MaterialIcons";
-import PrimaryButton from "@/components/PrimaryButton";
+import PrimaryButton from "@/components/common/PrimaryButton";
 import { AuthStackParamList } from "@/navigation/types";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
